Move Restriction, GuildBind, GroupBind, and EmbedPrompt to models.py
Circular imports are fun...
